requirements.txt
kafka-python==2.2.15
pyspark==3.3.0